Extract common code from experiment runners
